# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "User related user stories API."
  version: "0.1.9"
  title: "GreenHero API"
...
paths:
...
  /PersonalInfo:
    post:
      tags:
      - "PersonalInfo"
      summary: "Create a new personal Info"
      operationId: "createPersonalInfo"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new personal Info"
        required: true
        schema:
          $ref: "#/definitions/PersonalInfo"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "PersonalInfo"
  /PersonalInfo/{user_id}:
    get:
      tags:
      - "PersonalInfo"
      summary: "Find personal info of user's by user id"
      operationId: "getPersonalInfoByUserId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        description: "user id of personal info to return"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/PersonalInfo"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "PersonalInfo not found"
      x-swagger-router-controller: "PersonalInfo"
    put:
      tags:
      - "PersonalInfo"
      summary: "Modify PersonalInfo"
      description: "To modify a user's PersonalInfo "
      operationId: "modifyPersonalInfo"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        description: "The user id of the personal information that needs to be modified"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated personal info object"
        required: true
        schema:
          $ref: "#/definitions/PersonalInfo"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "Personal Info not found "
      x-swagger-router-controller: "PersonalInfo"
    delete:
      tags:
      - "PersonalInfo"
      summary: "Delete PersonalInfo"
      description: "To delete a user's personal info"
      operationId: "deletePersonalInfoByUserId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        description: "The user_id of the personal info that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "personal info not found"
      x-swagger-router-controller: "PersonalInfo"
...
  /Mission:
    get:
      tags:
      - "Mission"
      summary: "Get missions"
      description: "See the available missions"
      operationId: "getAllMissions"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Mission"
        "404":
          description: "No mission found"
      x-swagger-router-controller: "Mission"
    post:
      tags:
      - "Mission"
      summary: "Create a new mission"
      operationId: "createMission"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new mission"
        required: true
        schema:
          $ref: "#/definitions/Mission"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "Mission"
    put:
      tags:
      - "Mission"
      summary: "Modify mission"
      description: "To modify a mission "
      operationId: "modifyMission"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated mission object"
        required: true
        schema:
          $ref: "#/definitions/Mission"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "mission not found "
      x-swagger-router-controller: "Mission"
  /Mission/getMissionByTitleAndType:
    get:
      tags:
      - "Mission"
      summary: "Get mission"
      description: "See the available missions"
      operationId: "getMissionByTitleAndType"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "title"
        in: "query"
        description: "The title of the mission."
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "The type of the mission."
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Mission"
        "404":
          description: "No mission found"
      x-swagger-router-controller: "Mission"
  /Mission/{_id}:
    delete:
      tags:
      - "Mission"
      summary: "Delete mission"
      description: "To delete a mission"
      operationId: "deleteMissionById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "_id"
        in: "path"
        description: "The ID of the mission that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "mission not found"
      x-swagger-router-controller: "Mission"
...
  /SkillTree:
    post:
      tags:
      - "SkillTree"
      summary: "Create a new skill tree"
      operationId: "createSkillTree"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new skill tree"
        required: true
        schema:
          $ref: "#/definitions/SkillTree"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "SkillTree"
    put:
      tags:
      - "SkillTree"
      summary: "Modify skill tree"
      description: "To modify a skill tree"
      operationId: "modifySkillTree"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated skill tree object"
        required: true
        schema:
          $ref: "#/definitions/SkillTree"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "skill tree not found "
      x-swagger-router-controller: "SkillTree"
  /SkillTree/{_id}:
    get:
      tags:
      - "SkillTree"
      summary: "get SkillTree"
      description: "To get a skill tree by class"
      operationId: "getSkillTreeByClass"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "_id"
        in: "path"
        description: "The class of the skill tree"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/SkillTree"
        "400":
          description: "Invalid class supplied"
        "404":
          description: "SkillTree not found"
      x-swagger-router-controller: "SkillTree"
    delete:
      tags:
      - "SkillTree"
      summary: "Delete SkillTree by ID"
      description: "To delete a skill tree by ID"
      operationId: "deleteSkillTreeById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "_id"
        in: "path"
        description: "The ID of the skill tree that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "SkillTree not found"
      x-swagger-router-controller: "SkillTree"
...
  /Node:
    post:
      tags:
      - "Node"
      summary: "Create a new node"
      operationId: "createNode"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new node"
        required: true
        schema:
          $ref: "#/definitions/Node"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "Node"
    put:
      tags:
      - "Node"
      summary: "Modify node"
      description: "To modify a node"
      operationId: "modifyNode"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "modify the node"
        required: true
        schema:
          $ref: "#/definitions/Node"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "node not found"
      x-swagger-router-controller: "Node"
  /Node/{_id}:
    delete:
      tags:
      - "Node"
      summary: "Delete Node"
      description: "To delete a node"
      operationId: "deleteNode"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "_id"
        in: "path"
        description: "Id of node"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "No node found"
      x-swagger-router-controller: "Node"
  /Consumption:
    post:
      tags:
      - "Consumption"
      summary: "Create new Consumption"
      operationId: "createConsumption"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create new consumption"
        required: true
        schema:
          $ref: "#/definitions/Consumption"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "Consumption"
    put:
      tags:
      - "Consumption"
      summary: "Modify Consumption"
      description: "To modify a Consumption"
      operationId: "modifyConsumption"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "modify the Consumption"
        required: true
        schema:
          $ref: "#/definitions/Consumption"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "Consumption not found"
      x-swagger-router-controller: "Consumption"
  /Consumption/getConsumptionByUserIdandDate:
    get:
      tags:
      - "Consumption"
      summary: "Find consumption by user id and date"
      operationId: "getConsumptionByUserIdandDate"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "user id of consumption to return"
        required: true
        type: "string"
      - name: "date"
        in: "query"
        description: "date of consumption to return"
        required: true
        type: "string"
        format: "date"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Consumption"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Consumption not found"
      x-swagger-router-controller: "Consumption"
  /Consumption/deleteConsumptionByUserIdandDate:
    delete:
      tags:
      - "Consumption"
      summary: "Delete Consumption"
      description: "To delete a user's consumption of a date"
      operationId: "deleteConsumptionByUserIdandDate"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "The user id of the consumption that needs to be deleted"
        required: true
        type: "string"
      - name: "date"
        in: "query"
        description: "date of consumption"
        required: true
        type: "string"
        format: "date"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "User's consumption not found"
      x-swagger-router-controller: "Consumption"
  /Consumption/getConsumptionByUserIdDateAndCategory:
    get:
      tags:
      - "Consumption"
      summary: "Find consumption by user id,date and category"
      operationId: "getConsumptionByUserIdDateAndCategory"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "user id of consumption to return"
        required: true
        type: "string"
      - name: "date"
        in: "query"
        description: "date of consumption to return"
        required: true
        type: "string"
        format: "date"
      - name: "category"
        in: "query"
        description: "category  of consumption to return"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Consumption"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Consumption not found"
      x-swagger-router-controller: "Consumption"
  /Consumption/deleteConsumptionByUserIdDateAndCategory:
    delete:
      tags:
      - "Consumption"
      summary: "Delete Consumption by user id,date and category"
      description: "To delete a user's consumption by user id,date and category"
      operationId: "deleteConsumptionByUserIdDateAndCategory"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "The user id of the consumption that needs to be deleted"
        required: true
        type: "string"
      - name: "date"
        in: "query"
        description: "date of consumption"
        required: true
        type: "string"
        format: "date"
      - name: "category"
        in: "query"
        description: "category  of consumption"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "User's consumption not found"
      x-swagger-router-controller: "Consumption"
...
definitions:
...
  PersonalInfo:
    type: "object"
    properties:
      _id:
        type: "string"
      user_id:
        type: "string"
      realName:
        type: "string"
      email:
        type: "string"
        format: "email"
      occupation:
        type: "string"
      biography:
        type: "string"
      phone:
        type: "string"
      linkedin:
        type: "string"
      facebook:
        type: "string"
    xml:
      name: "personalInfo"
    example:
      realName: "realName"
      occupation: "occupation"
      user_id: "user_id"
      phone: "phone"
      facebook: "facebook"
      _id: "_id"
      biography: "biography"
      linkedin: "linkedin"
      email: "email"
...
  Mission:
    type: "object"
    required:
    - "title"
    properties:
      _id:
        type: "string"
      title:
        type: "string"
      description:
        type: "string"
      value:
        type: "number"
      goal:
        type: "number"
      reward:
        type: "integer"
      image:
        type: "string"
      deleted:
        type: "boolean"
      type:
        type: "string"
    xml:
      name: "Mission"
    example:
      reward: 1
      image: "image"
      goal: 6.02745618307040320615897144307382404804229736328125
      deleted: true
      description: "description"
      _id: "_id"
      title: "title"
      type: "type"
      value: 0.80082819046101150206595775671303272247314453125
...
  SkillTree:
    type: "object"
    properties:
      _id:
        type: "string"
      class:
        type: "string"
      nodes:
        type: "array"
        xml:
          name: "nodes"
          wrapped: true
        items:
          type: "string"
    xml:
      name: "SkillTree"
    example:
      nodes:
      - "nodes"
      - "nodes"
      _id: "_id"
      class: "class"
...
  Node:
    type: "object"
    properties:
      _id:
        type: "string"
      ability:
        type: "string"
      type:
        type: "string"
      amount:
        type: "number"
      augmentationType:
        type: "string"
      cost:
        type: "integer"
      sons:
        type: "array"
        xml:
          name: "sons"
          wrapped: true
        items:
          type: "string"
      locked:
        type: "boolean"
      owned:
        type: "boolean"
    xml:
      name: "Node"
  Consumption:
    type: "object"
    properties:
      _id:
        type: "string"
      user_id:
        type: "string"
      date:
        type: "string"
        format: "date"
      total:
        type: "number"
      category:
        type: "string"
    xml:
      name: "Consumption"
    example:
      date: "2000-01-23"
      total: 0.80082819046101150206595775671303272247314453125
      user_id: "user_id"
      _id: "_id"
      category: "category"
...